requirements.txt
pytest-playwright
pytest-html
pytest-clarity
